+++ rust-doomgeneric/src/m_fixed.rs
@@ -1,7 +1,7 @@
 pub type fixed_t = i32;
 pub const INT_MAX: i32 = i32::MAX;
 pub const INT_MIN: i32 = i32::MIN;
-pub const FRACBITS: i32 = 16 as i32;
+pub const FRACBITS: i32 = 16;
 pub fn FixedMul(a: fixed_t, b: fixed_t) -> fixed_t {
     return (a as i64 * b as i64 >> FRACBITS) as fixed_t;
 }
