+++ rust-doomgeneric/src/p_enemy.rs
@@ -90,9 +90,9 @@ pub const DI_NORTHWEST: dirtype_t = 3;
 pub const DI_NORTH: dirtype_t = 2;
 pub const DI_NORTHEAST: dirtype_t = 1;
 pub const DI_EAST: dirtype_t = 0;
-pub const ML_SOUNDBLOCK: i32 = 64 as i32;
-pub const MELEERANGE: i32 = 64 as i32 * FRACUNIT;
-pub const MISSILERANGE: i32 = 32 as i32
+pub const ML_SOUNDBLOCK: i32 = 64;
+pub const MELEERANGE: i32 = 64 * FRACUNIT;
+pub const MISSILERANGE: i32 = 32
     * 64 as i32 * FRACUNIT;
 #[no_mangle]
 pub static mut opposite: [dirtype_t; 9] = [
@@ -1188,7 +1188,7 @@ pub unsafe fn A_FatAttack3(mut actor: *mut mobj_t) {
     );
     (*mo).momy = FixedMul((*(*mo).info).speed as fixed_t, finesine[an as usize]);
 }
-pub const SKULLSPEED: i32 = 20 as i32 * FRACUNIT;
+pub const SKULLSPEED: i32 = 20 * FRACUNIT;
 pub unsafe fn A_SkullAttack(mut actor: *mut mobj_t) {
     let mut dest: *mut mobj_t = ::core::ptr::null_mut::<mobj_t>();
     let mut an: angle_t = 0;
