+++ rust-doomgeneric/src/p_pspr.rs
@@ -44,7 +44,7 @@ use crate::src::tables::FINEANGLES;
 use crate::src::p_enemy::MELEERANGE;
 
 
-pub const DEH_DEFAULT_BFG_CELLS_PER_SHOT: i32 = 40 as i32;
+pub const DEH_DEFAULT_BFG_CELLS_PER_SHOT: i32 = 40;
 pub const deh_bfg_cells_per_shot: i32 = DEH_DEFAULT_BFG_CELLS_PER_SHOT;
 pub unsafe fn P_SetPsprite(
     mut player: *mut player_t,
